+++ Accelerator/Exam/templates/Exam/questions.html
@@ -1,0 +1,15 @@
+{% extends 'Exam/base.html' %}
+{% block content %}
+<div class="card mb-4">
+                            <div class="card-body">
+                                <div class="table-responsive">
+                                    <table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
+                                                <th>Questions</th>
+{% for question in allQuestions %}
+                                        <tr><td>{{ question.qdesc }}</td></tr>
+{% endfor %}
+                                    </table>
+                                </div>
+                            </div>
+                        </div>
+{% endblock %}
