+++ Accelerator/Exam/urls.py
@@ -21,4 +21,5 @@ urlpatterns = [
     path('ExamHome', views.exam_home, name='exam-home'),
     path('StudentLogin/', views.student_login, name='student-login'),
     path('StudentLogout/', views.student_logout, name='student-logout'),
+    path('GetQuestions', views.getQuestions, name='getQuestions'),
 ]
